refactor: extract context query semantics
