+++ cereja/system/_context/__init__.py
@@ -1,0 +1,1 @@
+"""Private components for bounded textual context search."""
